+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.3] - [2019-10-02]
+
+### Added
+- Support of Rails 6 (@quikly)
+- Cached queries are ignored (@tyok)
+
 ## [0.0.2] - [2017-05-17]
 
 ### Added
@@ -16,3 +22,4 @@ This is a first public release (@nepalez)
 [Unreleased]: https://github.com/tram-rb/tram-policy
 [0.0.1]: https://github.com/nepalez/rspec-sqlimit/releases/tag/v0.0.1
 [0.0.2]: https://github.com/nepalez/rspec-sqlimit/compare/v0.0.1...v0.0.2
+[0.0.3]: https://github.com/nepalez/rspec-sqlimit/compare/v0.0.2...v0.0.3